accessible/tests/mochitest/states/test_docarticle.html

<html>
<head>
  <title>states of document article</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      var docAcc = getAccessible(document, [nsIAccessibleDocument]);
      if (docAcc) {
        testStates(docAcc, STATE_READONLY);
        testStates("article", STATE_READONLY);
        testStates("editable_article", 0, EXT_STATE_EDITABLE);

        document.designMode = "on";

        testStates(docAcc, 0, EXT_STATE_EDITABLE);
        testStates("article", 0, EXT_STATE_EDITABLE);
        testStates("article", 0, EXT_STATE_EDITABLE);
        testStates("editable_article", 0, EXT_STATE_EDITABLE);
  
        document.designMode = "off";

        testStates(docAcc, STATE_READONLY);
        testStates("article", STATE_READONLY);
        testStates("editable_article", 0, EXT_STATE_EDITABLE);
      }
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>

<body role="article">

  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=467387"
    title="Expose non-editable documents as readonly, regardless of role">
     Mozilla Bug 467387
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div id="article" role="article">article</div>
  <div id="editable_article" role="article" contentEditable="true">editable article</doc>
</body>
</html>
